# backend/src/swagger/schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Easy-Quizi API",
    "description": "API definition",
    "termsOfService": "http://swagger.io/terms/",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "User",
      "description": "Operations about user"
    },
    {
      "name": "Quiz",
      "description": "Operations about quiz"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logs user into the system",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsrLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout user from the system",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["User"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": ["User"],
        "summary": "Create new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes": {
      "get": {
        "tags": ["Quiz"],
        "summary": "Get all quizzes",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quiz"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/create": {
      "post": {
        "tags": ["Quiz"],
        "summary": "Create new quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizBase"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quiz"
                }
              }
            }
          }
        }
      }
    },
    "quizzes/result/send": {
      "post": {
        "tags": ["Quiz"],
        "summary": "Send result",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Quiz"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Score"
                }
              }
            }
          }
        }
      }
    },
    "/quizzes/delete/{id}": {
      "delete": {
        "tags": ["Quiz"],
        "summary": "Delete quiz by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["fullName", "email", "password", "status"],
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" },
          "fullName": { "type": "string" },
          "status": { "type": "string" }
        },
        "example": {
          "fullName": "Obi Wan Kenobi",
          "email": "ObiWanKenobi@easy-quizi.com",
          "status": "teacher",
          "password": "123456"
        }
      },
      "UsrLoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        },
        "example": {
          "email": "ObiWanKenobi@easy-quizi.com",
          "password": "123456"
        }
      },
      "Question": {
        "type": "object",
        "required": ["question", "answers", "rightAnswers"],
        "properties": {
          "uid": { "type": "string" },
          "question": { "type": "string" },
          "answers": {
            "type": "object",
            "additionalProperties": {}
          },
          "rightAnswers": { "type": "array", "items": { "type": "string" } }
        }
      },
      "QuizBase": {
        "type": "object",
        "required": ["title", "questions"],
        "properties": {
          "title": { "type": "string" },
          "time": { "type": "string" },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        },
        "example": {
          "title": "Web developing UI/UX 1",
          "time": "25",
          "questions": [
            {
              "question": "Why UI/UX so important?",
              "answers": {
                "0": "because",
                "1": "it is important",
                "2": "not important",
                "3": "I do not know"
              },
              "rightAnswers": [0, 1]
            },
            {
              "question": "Why are u gay?",
              "answers": {
                "0": "because",
                "1": "I am not gay"
              },
              "rightAnswers": [0]
            }
          ]
        }
      },
      "Quiz": {
        "allOf": [
          {
            "type": "object",
            "required": ["uid"],
            "properties": {
              "uid": { "type": "string" },
              "createdBy": { "type": "string" }
            }
          },
          { "$ref": "#/components/schemas/QuizBase" }
        ]
      },
      "Score": {
        "type": "object",
        "properties": {
          "score": { "type": "string" }
        }
      }
    }
  }
}